#53 payment system: payment service - amount 타입 변경(BigDecimal -> Long)

+++ payment-system/paymentservice/src/main/kotlin/org/example/paymentservice/payment/domain/PaymentOrder.kt
@@ -1,14 +1,12 @@
 package org.example.paymentservice.payment.domain
 
-import java.math.BigDecimal
-
 data class PaymentOrder(
     val id: Long? = null,
     val paymentEventId: Long? = null,
     val sellerId: Long,
     val productId: Long,
     val orderId: String,
-    val amount: BigDecimal,
+    val amount: Long,
     val paymentStatus: PaymentStatus,
     private val isLedgerUpdated: Boolean = false,
     private val isWalletUpdated: Boolean = false,
