+++ payment-system/paymentservice/src/main/kotlin/org/example/paymentservice/payment/domain/Product.kt
@@ -1,10 +1,8 @@
 package org.example.paymentservice.payment.domain
 
-import java.math.BigDecimal
-
 data class Product(
     val id: Long,
-    val amount: BigDecimal,
+    val amount: Long,
     val quantity: Int,
     val name: String,
     val sellerId: Long,
